Change price data type to float in struct def

+++ header.h
@@ -1,8 +1,7 @@
 typedef struct {
-
     int id;                 //Item ID
     char description[50];   //Item Description
     int qty;                //Item Quantity
     char exp[10];           //Item Expiry Date
-    double price;           //Item price
-} item;
+    float price;           //Item price
+  } item;
